move Configuration Impl closer to point of definition, isolating code from the sudoers crate further
