+++ lib/sudo-env/src/environment.rs
@@ -1,7 +1,8 @@
 use std::collections::HashSet;
 
-use sudo_common::context::{CommandAndArguments, Configuration, Context, Environment};
+use sudo_common::context::{CommandAndArguments, Context, Environment};
 use sudo_system::PATH_MAX;
+use sudoers::Configuration;
 
 use crate::wildcard_match::wildcard_match;
